font-patcher: Link to icon sources in patched fonts
[why]
Some icon sets require mentioning of the original author, and we do this
on the website. But there is no indication that we have that list.

[how]
Remove the comment about the changelog (usually you look at the
changelog at the repo before you install a new font) and substitute it
for a link to the icon repo list.

Related: #1908

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>